+++ app/Partner.php
@@ -2,7 +2,9 @@
 
 namespace App;
 
+use DateTime;
 use Illuminate\Support\Str;
+use Illuminate\Support\Carbon;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
 
@@ -12,6 +14,17 @@ use Illuminate\Database\Eloquent\Builder;
 class Partner extends Model
 {
     use HasContactDetails;
+
+    /**
+     * The attributes that should be mutated to dates.
+     *
+     * @var array
+     */
+    protected $dates = [
+        'joined_on',
+        'left_on',
+        'validated_at',
+    ];
 
     /**
      * The "booting" method of the model.
@@ -35,6 +48,13 @@ class Partner extends Model
         // the currency, because most of the time we won’t want them.
         static::addGlobalScope('active', function (Builder $builder) {
             $builder->whereNull('left_on');
+        });
+
+        // Add a default global scope to all select queries on the model.
+        // This will exclude nonvalidated partners, who have not been
+        // accepted (yet) into the network.
+        static::addGlobalScope('validated', function (Builder $builder) {
+            $builder->whereNotNull('validated_at');
         });
     }
 
@@ -75,7 +95,41 @@ class Partner extends Model
     }
 
     /**
+     * Check if the partner has been validated or not.
+     *
+     * @return bool
+     */
+    public function isValidated()
+    {
+        return $this->validated_at instanceof DateTime;
+    }
+
+    /**
+     * Mark the partner as valid from now.
+     *
+     * @return void
+     */
+    public function validate()
+    {
+        $this->validated_at = Carbon::now();
+        $this->save();
+    }
+
+    /**
+     * Invalidate the partner.
+     *
+     * @return void
+     */
+    public function invalidate()
+    {
+        $this->validated_at = null;
+        $this->save();
+    }
+
+    /**
      * Return the list of cities from the address(es) of the partner’s locations.
+     *
+     * Cities are sorted in alphabetical order.
      *
      * @return string|null
      */
@@ -97,6 +151,9 @@ class Partner extends Model
         // Remove duplicates in case there are multiple
         // locations in the same city.
         $cities = array_unique($cities);
+
+        // Then, sort the cities in alphabetical order.
+        sort($cities, SORT_LOCALE_STRING);
 
         return $cities ? implode(', ', $cities) : null;
     }