+++ CHANGELOG.md
@@ -7,10 +7,16 @@ The format is based on
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html "Semantic Versioning 2.0.0 specification").
 
-## [Unreleased]
+## [1.1.0] - 2017-10-31
+### Added
+- Partners now have a validation state. They can be valid or not.
+- Improves performance when retrieving the list of partners.
 
-## [1.0.0] - 2017-10-26
+## 1.0.0 - 2017-10-26
 ### Added
 - Home page of the site.
 - A page listing existing currency exchanges (hardcoded).
 - A page listing active partners of the currency (dynamic).
+
+[Unreleased]: https://github.com/monnaie-valheureux/radisse/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/monnaie-valheureux/radisse/compare/v1.0.0...v1.1.0